Move RPC method
Moved the RPC method for route recalculation from /api/routecalc to /api/area.  This is a more logical place as the old method required you to know the routecalc_id and the area_id while the new method is only the area_id
