+++ madmin/templates/settings_areas.html
@@ -42,13 +42,10 @@ $(document).ready(function () {
             var elem =  $(this);
             $.blockUI({ message: '<img src="{{ url_for('static', filename='loading.gif') }}" width="100px" /><br /><h2>Recalculating routefile...</h2>' });
             rpc_call = {
-              'call': 'recalculate',
-              'args': {
-                'area_id': $(this).data('area')
-              }
+              'call': 'recalculate'
             }
             $.ajax({
-                url : '{{ url_for('api_routecalc') }}/'+ $(this).data('identifier'),
+                url : '{{ url_for('api_area') }}/'+ $(this).data('area'),
                 contentType : 'application/json-rpc',
                 data: JSON.stringify(rpc_call),
                 type : 'POST',
@@ -141,7 +138,7 @@ $(document).ready(function () {
             </p>
           </td>
           <td class="text-right align-middle">
-            <button data-identifier='{{ area['routecalc'] }}' data-area='{{ area_id }}' type="button" class="recalculate btn btn-info btn-sm"><i class="fa fa-sync"></i></button>
+            <button data-area='{{ area_id }}' type="button" class="recalculate btn btn-info btn-sm"><i class="fa fa-sync"></i></button>
             <a href="{{ redirect }}?id={{ area_id }}&mode={{ area.area_type }}"><button type="button" class="btn btn-success btn-sm edit" data-identifier="{{ loop.index }}"><i class="far fa-edit"></i></button></a>
             <button data-identifier='{{ area_id }}' type="button" class="delete btn btn-danger btn-sm"><i class="fas fa-trash-alt"></i>
           </td>
